projeto_etl_dados/requirements.txt
pandas==2.1.4
openpyxl==3.1.2
sqlalchemy==2.0.25
pydantic==2.5.3
psycopg2-binary==2.9.9
geopy==2.4.1
python-dotenv==1.0.0
pytest==7.4.3